refactor: rename abstract trainer folder and move some files
