+++ training-yolov8-classification/pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 description = ""
 authors = ["hajer <hajer.essefi@picsellia.com>"]
 readme = "README.md"
-packages = [{include = "training-yolov8-classification"}]
+packages = [{ include = "training-yolov8-classification" }]
 
 [tool.poetry.dependencies]
 python = "^3.11"
@@ -13,6 +13,7 @@ ultralytics = "8.0.67"
 onnx = "1.13.1"
 scikit-learn = "1.2.2"
 coverage = "7.3.0"
+picsellia = "6.10.0"
 
 [tool.poetry.group.dev.dependencies]
 coverage = "^7.3.0"
